Refactor: Refactors fees to use custom Fee entity [phamhminhkhoi]
Replaces the String list for fees in Quotation entity with a custom Fee entity to store fee name, amount, and currency.
Introduces a FeeListConverter to handle the conversion between the Fee list and a string representation in the database.
Updates the OnlineQuotationDTO and OnlineQuotationRequest to use the Fee entity and FeeRequest object respectively.

This change allows for more structured and type-safe handling of fees within the application.